Updated version code Theme fixes for the alert dialog Code refactor
