+++ app/src/main/java/com/neuralbit/letsnote/firebaseRepos/DeleteDataRepo.kt
@@ -1,4 +1,4 @@
-package com.neuralbit.letsnote.repos
+package com.neuralbit.letsnote.firebaseRepos
 
 import android.content.Context
 import android.content.Intent
