+++ app/src/main/java/com/neuralbit/letsnote/firebaseRepos/NoteFireRepo.kt
@@ -1,4 +1,4 @@
-package com.neuralbit.letsnote.repos
+package com.neuralbit.letsnote.firebaseRepos
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
