+++ app/src/main/java/com/neuralbit/letsnote/ui/label/LabelViewModel.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import com.neuralbit.letsnote.entities.LabelFire
-import com.neuralbit.letsnote.repos.LabelFireRepo
+import com.neuralbit.letsnote.firebaseRepos.LabelFireRepo
 
 class LabelViewModel (application: Application): AndroidViewModel(application)  {
     var labelList: ArrayList<Label>  = ArrayList()
