+++ app/src/main/java/com/neuralbit/letsnote/utilities/RescheduleAlarmsService.kt
@@ -9,7 +9,7 @@ import android.content.Context
 import android.content.Context.ALARM_SERVICE
 import android.content.Intent
 import com.neuralbit.letsnote.entities.NoteFire
-import com.neuralbit.letsnote.repos.NoteFireRepo
+import com.neuralbit.letsnote.firebaseRepos.NoteFireRepo
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
