+++ app/src/main/res/layout/alert_datetime_dialog.xml
@@ -1,13 +1,16 @@
 
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
 xmlns:tools="http://schemas.android.com/tools"
-android:layout_width="300dp"
-xmlns:app="http://schemas.android.com/apk/res-auto"
-android:padding="16dp"
-android:id="@+id/releaseNotesLayout"
+    android:layout_width="300dp"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:padding="16dp"
 
-android:background="@drawable/save_shared_background"
-android:layout_height="wrap_content">
+    android:id="@+id/releaseNotesLayout"
+
+    android:background="@drawable/save_shared_background"
+    android:layout_height="wrap_content"
+    android:theme="@style/Custom_dialog_style"
+    >
 
     <TextView
         android:id="@+id/reminderTitleTv"
@@ -27,6 +30,7 @@ android:layout_height="wrap_content">
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:text="@string/choose_date"
+        android:textColor="@color/gunmetal"
         android:textSize="16sp"
         android:layout_marginTop="8dp"
         app:layout_constraintTop_toBottomOf="@id/reminderTitleTv"
@@ -40,6 +44,7 @@ android:layout_height="wrap_content">
         android:background="?attr/selectableItemBackgroundBorderless"
         android:contentDescription="@string/date_button"
         android:src="@drawable/ic_baseline_date_range_24"
+        android:textColor="@color/gunmetal"
         app:layout_constraintBottom_toBottomOf="@+id/dateTitle"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="@+id/dateTitle" />
